refactor: partially decouple sidebar from BookPage
