+++ src/app/components/Sidebar/index.tsx
@@ -6,7 +6,7 @@ import LIBMSLOGO from './assets/LibMSLOGO.png';
 
 const navItemsTop = [
   { name: 'Dashboard', link: '/dashboard', icon: IconName.FaChartBar },
-  { name: 'Library', link: '/books', icon: IconName.FaBook },
+  { name: 'Library', link: '/dashboard/books', icon: IconName.FaBook },
   { name: 'Reservations', link: '/dashboard', icon: IconName.FaRegListAlt },
   { name: 'Templates', link: '/dashboard', icon: IconName.FaColumns },
   { name: 'Categories', link: '/dashboard', icon: IconName.FaTag },
