+++ src/app/index.tsx
@@ -9,7 +9,7 @@
 import * as React from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useTranslation } from 'react-i18next';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Switch } from 'react-router-dom';
 import { createGlobalStyle } from 'styled-components';
 import { NotFoundPage } from './components/NotFoundPage/Loadable';
 import { AuthenticatedRoute, UnauthenticatedRoute } from './guards/Routes';
@@ -17,7 +17,7 @@ import { AddBookPage } from './pages/AddBookPage/Loadable';
 import { AuthPage } from './pages/AuthPage/Loadable';
 import { BookDescPage } from './pages/BookDescPage';
 import { BooksPage } from './pages/BooksPage/Loadable';
-import { DashboardPage } from './pages/DashboardPage/Loadable';
+import { DashboardPage } from './pages/DashboardPage';
 import { HomePage } from './pages/HomePage/Loadable';
 import { LoginPage } from './pages/LoginPage/Loadable';
 import { RegisterPage } from './pages/RegisterPage/Loadable';
@@ -53,10 +53,10 @@ export function App() {
 
         <AuthenticatedRoute exact path="/user" component={UserProfilePage} />
         <AuthenticatedRoute exact path="/add-book" component={AddBookPage} />
-        <AuthenticatedRoute exact path="/dashboard" component={DashboardPage} />
+        <AuthenticatedRoute path="/dashboard" component={DashboardPage} />
         <AuthenticatedRoute exact path="/books" component={BooksPage} />
         <AuthenticatedRoute exact path="/books/:id" component={BookDescPage} />
-        <Route component={NotFoundPage} />
+        <AuthenticatedRoute component={NotFoundPage} />
       </Switch>
       <GlobalStyle></GlobalStyle>
     </BrowserRouter>
